+++ src/test/elements/googlecalendar/calendars.js
@@ -4,6 +4,6 @@ const suite = require('core/suite');
 const tools = require('core/tools');
 const payload = tools.requirePayload(`${__dirname}/assets/calendars.json`);
 
-suite.forElement('general', 'calendars', { payload: payload }, (test) => {
+suite.forElement('scheduling', 'calendars', { payload: payload }, (test) => {
   test.should.supportCrud();
 });
